VK_RIGHT now has no error but doesn't move still
Garrett changed where you make the laserCannonPic (he removed the class
name LaserCannon before the variable name laserCannonPic). This made
the
Point location = laserCannonPic.getLocation();
not throw an error and it now prints all the test prints in the
VK_RIGHT spot. Now we just need to make it actually move and repaint.