Read description
Changed variable name to be more descriptive.
Changed collide() to throw an alien off of the screen after getting hit
and added a safeguard to prevent the alien movement rotation from
triggering and rapidly throwing the remaining aliens forward and to the
side.
Corrected issue in which alien shots were not being removed from the
game.
Corrected issue in which player shots would not automatically be removed
from the game following a collision.