Respawn point, laserCannon bound.
Changed the respawn point such that if a player is killed by colliding
with an alien ship they will spawn back to a place aliens will never be.
Added bounds such that the laser cannon can not leave the screen.
Messed around with all kinds of Eclipse source tools and probably broke
things.